Create Top Movie Genres by Country
This query was written to find the top movie genres from each country where a movie rental company is doing business. It pulls the number of customers in each country, the sum of money spent by those customers, and the most rented movie genre for the top 10 countries ranked by total money spent.